refactor: deduplicate TypeScript-based code
