+++ src/rules/utils/index.ts
@@ -3,3 +3,4 @@ export * from './detectJestVersion';
 export * from './followTypeAssertionChain';
 export * from './misc';
 export * from './parseJestFnCall';
+export * from './ts';
